refactor specific rfc and split discussion tab
